pyproject.toml
[tool.poetry]
name = "yaschedule-replit"
version = "0.1.0"
description = ""
authors = ["Pavel <pavel.p.serikov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
yaschedule = {git = "https://github.com/StannisGr/yaschedule.git"}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
